pyproject.toml
[tool.poetry]
name = "melnor-bt"
version = "0.1.0"
description = "A small python library for discovery and interacting with Melnor, Aquastar, etc bluetooth water timers."
authors = ["Justin Vanderhooft <justinvdhooft@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bluepy = "^1.3.0"
"backports.zoneinfo" = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
